Separated layout from index.js.
